.emacs.el: smoother-kill-buffer: robustness++
If some other kill buffer query function or
kill buffer hook runs before ours and causes
the buffer to no longer be shown, our hook's
`quit-window` call would do the wrong thing.

One indirect example of this is the Backtrace
buffer: the debug system uses `recursive-edit`,
so the Backtrace buffer calls `top-level` when
being killed during a recursive edit. This in
turn *prevents* the Backtrace buffer from being
killed, because aborting a recursive edit will
abort the current command in that edit's
command loop. To actually force the buffer to
be killed as intended, the Backtrace buffer
needs a modified kill-buffer-hook which (in
addition to aborting recursive-edit) registers
a post-command-hook to killing itself (which
will fire once the recursive-edit is aborted).
By the time this post-command-hook fires, the
the buffer is no longer being displayed (this
may be either due to our kill hook already
running once in the recursive edit, or due to
the `save-excursion` in `debug` that wraps the
`recursive-edit` call).

Also, other hooks changing windows can cause
the opposite problem: the buffer being killed
might be displayed in no windows at the start
of the kill, but then get brought into view
before our hook runs.

More generally, `(quit-window)` is wrong here in
the general case even without window changes: if
the buffer is being displayed in any window(s)
other than the selected one, it would be killed
without quitting, and the not-targeted selected
window would quit instead.

To fix all of these problems, our hook is now
always added, and calls `quit-window` on each
window that is displaying the targetted buffer.