.emacs.el: make Backtrace quit not overreach
`debugger-quit` invokes `top-level`, which is
the cause of a rather frustrating issue where
I'll be using recursive minibuffers, and an
error in an innermost eval will wipe out the
outer ones (which rather defeats the purpose
of recursive minibuffers - the whole point is
to save time/effort by not having to manually
re-enter my current minibuffer contents after
a side-track, but with this debugger "feature"
I end up having to do just that).

This commit fixes that by aborting just one
level of minibuffer instead.