Revert ".emacs.el: make Backtrace quit not overreach (v2)"
This reverts commit 400dc5f16e0781dce28b4b39ea61e0a1da18f7ce.

Looking at Emacs' debugger more closely, I
"get it" now: `debug.el` is not wrong for
putting the recursive-edit abort into a
special quit function: you could validly
want to exit from the debugger in other
ways - for example, by continuing execution.

I was biased/tunnel-visioned because most of
my experience with the built-in debugger is
as a mere traceback on errors from manual
evals, where continue and abort are basically
the same (both immediately bubble up and
error out).

In other words, `debugger-quit` makes some
sense as a harder "abort" function, rather
than as the default window quit behavior.

(The thing that got me thinking about this
is that is doesn't make sense for quitting
to do more "exiting" than other ways of
putting away the buffer/window.)