.emacs.el: fix Backtrace kill recursion edge-case
This commit ensures that the debugger's recursive
edits are aborted before killing the buffer. This
fixes an edge case where a suspended minibuffer
that was started inside a debugger's recursive
edit would cause an error (`read-from-minibuffer`
would attempt to switch back to the backtrace
buffer while handling the thrown exit).