bin/gp: remove unused corner-case guard
It looks like I was intending to do this, but
decided against it or forgot or whatever.

The idea was to protect against the extremely
narrow race where a signal arrives between the
shell associating (enough of) the output from
`mktemp` with `$directory`, and finishing the
`trap` call.

But on further thought, I technically can't
be sure that the shell wouldn't do a partial
copy of `mktemp` output to whatever memory is
already associated with `directory`, and in
turn that partial string could in theory be
another path that the caller has permission
to `rm -rf`. Extremely unlikely, but I won't
trade potentially unboundedly destructive
misbehavior for a miniscule tightening of
cleanup likelihood.