.emacs.el: sorted-hash-table-put: try both ends
Search both ends of the dlist when adding a new
item. One big reason why I kept add-to-back
code outside of sorted-hash-table is that I
didn't like muddling an API which fits all
sorted collection implementations with a detail
that only fits this specific implementation in
which adding is O(1) at the tips and O(n) in
the middle. This new approach doesn't have that
problem - the API is still just an abstract
insertion, and callers automagically get O(1)
for inserts at either edge without having to
know anything.

While I'm at it, refactor to use records.