.emacs.el: insert-vertico-candidate-for-crm: keep query
Often, when I'm using completing-read-multiple
and I pick one candidate, I want to pick other
candidated from the same narrowed list, or
use a modified query.

I originally thought negation would be good
enough for this use-case, but it turns out the
UX doesn't really work - it's mentally out of
the way and more expensive. (Natural cognition
is NOT layering on one or more passes of
not-what-I-want markers and then grabbing all
items left unmarked. When we open a CRM prompt
and narrow, that's like walking into a store
or library, and finding the right shelf. The
natural UX is that when you pick up the one
thing you know you want off the shelf, you can
turn your attention back to the shelf - the
shelf hasn't been rearranged and you have not
been teleported back to the front door.)