.emacs.el: insert-vertico-candidate-for-crm: don't lose place
This commit further enhances my integration of
`completing-read-multiple` and Vertico, so that
"," doesn't lose the current position in the
candidate list.

Same as the previous improvement: my gaze
doesn't involuntarily jump to the start of a
shelf when I pick up one item from the shelf,
and multi-selecting UIs shouldn't either.