bin/git-pop: a `git stash pop -p` UX
This implements an interactive, hunk-based
stash pop - same as the `-p` option of Git
commands like `git add`, `git checkout`,
and so on.

(Sadly, because this is a quick hack from
outside of Git, it leaves the repo in an
~intermediate state, with temporary commits
and a mutated index, until it's finished.
I'm sure there's some way using plumbing
commands or just a lot more code+effort to
do this a lot better, but I just want an
MVP that gets me the human-level practical
functionality.)