.emacs.el: histdir: don't falsely hold buffer-local latest entry
1. When buffer is first created, or

2. when the local latest entry is deleted,

there is no local latest entry. Now instead
of pinning the latest synced entry as the
latest entry in all cases, the code supports
not having a pinned latest entry.

(This makes a difference when watching sync
or waiting for a specific entry to appear;
it's less ergonomic and potentially slightly
confusing to have some entry you never ran
in the current shell pinned as the latest.)