.emacs.el: histdir: ignore wrongly-named files
This reduces in-memory history churn from the
coming and going of temporary files. For
example, Syncthing initially saves files as
`.syncthing*` in the same directory, until it
finishes writing their content. (Many other
programs similarly save files to temporary
names in the same directory. This is the
standard technique for making edits/creation
of a file at a given path effectively atomic.
For example, typical `sed` implementations
do this.)