.emacs.el: histdir: fix sync/watch when call file appears before string file
File sync can happen out-of-order: a call file
can be created after the string file that it
points to, but then end up finishing sync to
another device in the opposite order.

This commit watches the string directory along
with the call directory: if the call watcher
sees a new call file with no string file, it
now records the hash->datetime association,
rather than totally ignoring the file; and the
new string watcher adds the datetime->string
part for for any hash->datetime associations.
(So if they sync in the intended order, it all
works as before: the string watcher doesn't yet
have any partial entries to update, and then
the call watcher does the full entry add.)

P.S. This commit is a great example of why
idempotent operations are the right design
in most cases, because of how much simpler
and clearer (and easier to implement) they
make implementing things of non-trivial
complexity.