.emacs.el: vi-style delete in Dired
Big picture, this is one piece of a better
(and more vi-like) UX for file move and copy.

Note:

1. vi's `d` deletes from the file *in the
   editor*, but not from the file on disk
   - modification of file on disk is a
   separate save operation.

2. vi's `d` is actually a *cut* operation
   - the text is still in the clipboard,
   not totally gone.

So the true analog of vi's `d` for file
management is actually a cut operation,
which only deletes the file's line *in the
file manager* (and annotates it so that
paste commands can treat it as a move).

This commit implements `evil-dired-delete`
which does that, re-using my `full-path`
property and adding `paste-is-move`.