.emacs.el: move git-reset to " vh" (from " vj")
Turns out the "j" is so totally unintuitive to
me that I had to try almost every unbound
letter before I found it. Trying "h" for "head"
now, since Git's reset is a moving of the HEAD
(possibly with index chqnges, or both index
and worktree changes), which will hopefully be
more mnemonic/memorable/intuitive/automatic.