.emacs.el: independent-dired: tighter Dired integration
This commit makes `independent-dired` buffers
show up in `dired-buffers` (and advises
`dired-find-buffer-nocreate` to omit buffers
created by independent-dired).

As a result, most Dired functions (for example
`dired-{add,remove,relist}-file`, and anything
else based on `dired-fun-in-all-buffers` or
`dired-buffers-for-dir[-or-subdir]`) will now
affect buffers created with `independent-dired`
(but those buffers continue to not participate
in Dired's normal buffer reuse).