.emacs.el: enhance Evil's linewise yanking
By default, `evil-yank-lines` injects a trailing
newline if the yanked text doesn't have one. But
this is wrong for normally-single-line use-cases
such as the minibuffer and Eshell, where the
input is conceptually normally a field rather
than a block.

This commit adds a buffer-local variable which
allows disabling this behavior for single-line
linewise yanks which yank less than a full line.