.emacs.el: window-state: enhance "send" operator
The send operator is special among the non-"move"
operators: its selection of the target window
is user-visible. So it should record its switch
to the target window. (As a practical matter,
this results in correctly moving the "brighter"
background of the active/selected window to the
target window of "send", so that the UI reflects
where the next user input is going to go.)