.emacs.el: shred buffer
First pass at implementing the concept of a
generic buffer that deletes/destroys things.

The general idea is that you can delete or
clean up things by dropping them into this
buffer. (Only by pasting for now, although
for full feel of the vision, also imagine
full integration with system drag-and-drop.)

The boring foundation of this is pasting
to pop from Emacs' kill-ring, or pasting
from registers to clear them. The key is
that text can be annotated with the
`shred-handler` property to add additional
shredding behavior. For example, a file
path could be annotated with a handler
that asks for confirmation and then calls
the `shred` program (or `rm`, or `trash`,
or ...) on that file.