.emacs.el: dired-filename-search-forward: tolerate broken Dired buffers
A normal Dired buffer always has at least one
line, and the last line is always terminated
by a newline (linefeed) character.

`dired-filename-search-forward` has a loop at
the beginning to skip over non-entry lines,
such as (sub)directory headers. It would loop
forever if the last line didn't end in a
newline and wasn't a directory entry.

In practice, I have a few reasons for
deleting the top line (directory path / header)
of some Dired buffers:

1. It doesn't cause buffer undo/redo to trip
   over a full buffer revert the way that
   widen-and-narrow-again do. (Came up recently
   as I'm trying to finish off some vi-style
   controls for Dired buffers.)

2. It's faster (sometimes to a human-perceptible
   worse-UX degree), when heavy fontification is
   in play (that was the original reason I tried
   deleting rather than narrowing).

3. Ace Window's overlays don't handle narrowing
   correctly (this needs to be fixed generally,
   but for now the main (only?) place I observe
   it is when one of my Task buffers is empty).

And once deleting the top line is in play, it
becomes possible for the Dired buffer to be
totally empty, and then this infinite loop bug
comes up When I do most of my window
management/switching, because Ace Window
temporarily adds spaces to empty buffers to
make its overlays work consistently, and then
font-lock runs on redisplay, and I use
`dired-filename-search-forward` in my Denoted
naming fontification.

So this commit fixes the loop condition to
not get stuck in an infinite loop. (It does
NOT enhance this function in any other way:
in particular, the return value if a Dired
buffer contains only spaces can now be a junk
match (should be nil), and an error is thrown
if the buffer ends with a Dired entry that for
some reason is missing its trailing newline
(should ideally return the same result as it
would if the trailing newline was present).)