.emacs.el: form-delete
This allows removing forms rather than
just replacing them with other forms.

`string-replace` doesn't need an analog
to this because strings are flat - in
flat sequences, replacing with empty
is the same as removing. But structured
things that can nest have a distinction
that must be explicity disambiguated:
replace a node with `nil`, or delete
the node entirely.