.emacs.el: form-replace: n-to-n generalization
The last commit allowed replacing with
sequences of forms. This allows matching
sequences of forms to be replaced.

Besides making it possible to replace two
or more forms that are not conviently
in a small larger form (and which must be
matched together to avoid false positives),
this brings the interface back to the
idiomatic, symmetric, and nice-for-Lisp
API shape of `string-replace`.