.emacs.el: function-lisp-anonymize
This function homogenizes Lisp function
definitions (such as from `defun`) into
their `lambda` (or closure) forms.

This pairs well with `function-lisp`, which
might need to fall back on reading the form
which defined the function from source, and
thus return a form that needs to be handled
differently based on entirely orthogonal
concerns such as whether the function has
been compiled.

Now code using `form-replace` to define
patched functions can just use

    (function-lisp-anonymize
        (function-lisp ...))

to decouple itself from all that.

(Eventually, this function should be extended
to try matching the definition form against
entries in a publicly exposed list, so that
adding support for function-defining macros
can be done from outside the function, but
this MVP is good enough for now.)