.emacs.el: hack to make hack work with hack
Clearing file name handler during init is
nice way to shave some time off init (about
0.2 seconds in my case on my phones, which
is just long enough for me to feel).

But we end up needing a handler for ".gz"
if my `function-lisp` needs to fall back
on `find-function-*` to read the source,
and the source is one of the files that
ships gzipped.

Which comes up in my other hack that makes
Dired preserve buffer undo.

(Too many hacks, and opening `dired.el.gz`
often adds around 0.1 seconds of delay to
my init (or, rarely, a few times longer),
so I might eventually rewrite `dired-readin`
in my config instead of all this, but this
is the fastest path to what I need for now.)