.emacs.el: fixed-{start,end}-of-paragraph-text
When I'm at the start of a paragraph, and
I ask to be at the start of the current
paragraph, the only correct result is for
me to stay where I am, not be move to the
start of the previous paragraph.

It is damning to the existing factoring/APIs
in `paragraphs.el` that this is the easiest,
simplest, and cleanest way to reuse existing
paragraph edge-finding/identifying logic.