.emacs.el: smoother-fill-paragraph-post-command (draft)
Makes a mess of the undo history. Sometimes,
hitting Undo/Redo or trying to open Undo Tree
hangs and needs a Ctrl+G (maybe linked to Undo
right after Evil delete commands). And this is
a permanent/lingering condition - undo for that
buffer seems to get poisoned forever. When it's
not breaking undo entirely, it's adding a lot
of no-op entries.

Also, since it's constantly rewriting what
you just changed, it sometimes blocks certain
actions:

* backspacing at the left edge of an
  indented paragraph is sometimes a
  no-op, like when point is on this line;

* creating new paragraphs the normal way
  (hitting Enter) is basically always a no-op.

But it's good enough that I can leave it running
most of the time while writing the body of these
commit messages. (I do find myself toggling it
on-and-off a lot, but at least half the time
while writing this message it has stayed on.)