.emacs.el: fringe-like line-wrap indicator
Emacs in a terminal (non-graphic frames) lacks
fringes, which means line wrap indicators are
either worse (`\` at the *end* of lines) or
missing (when visual-line-mode is turned on).

This commit fakes the left fringe, using
`line-prefix` and `wrap-prefix`.

(For now I am limiting this feature to Termux.
The ideal is for this to always happen on every
non-graphical frame and to never happen on
graphical frames, but there doesn't seem to be
a way to do that within the same Emacs server
instance. Maybe one day I'll figure out some
way to do it, perhaps with overlays, display
properties, face remapping, or something like
that. but for now this is good enough. I
approximately never use Emacs in a terminal
outside of Termux, and can't have graphical
frames on Termux.)