.emacs.el: with-variable-watcher
In the same vein as my `with-{advice,hook}`,
for variable watchers.

Compliments `with-advice` in particular for
the kijd of extensive monkey-patching I do,
since byte-compiled functions don't bypass
variable watchers the way they bypass advice
in calls to other compiled functions.

(Also, big moment for me: I am allowing this
function to break the 80 character limit, which
I have religiously stuck to for close to a
decade now. Because in this case fitting in
less while maintaining all my other criteria is
non-trivial, and god damnit I want to LIVE.)