.emacs.el: smoother-eval-last-sexp
Like `eval-last-sexp`, but in the unamgiguous
case where it is called on the starting
parenthesis of an s-expression, it evaluates
that s-expression instead of erroring out with
a "Containing expression ends prematurely"
scan-error.

Was putting this one off for a year or two
but it has been coming up a lot lately.