.emacs.el: read-buffer-multiple: show default
Originally didn't bother with this because in
the common case the default buffer is already
the first Vertico candidate, and I think it's
actually better/non-redundant UX to just have
the default indicated that way.

But hidden buffers aren't shown by default
as candidates (f.e. when running
`smoother-kill-buffer` with a hidden buffer
as current), resulting in misleading UI
cues (I just found myself doubting/worrying
if it would do the right thing while
killing to leave a hidden buffer).