.emacs.el: command-execute-in-keymap: fix `{this,last}-command`
This ensures that the executed command sees
itself in `this-command` and the next command
sees the executed command in `last-command`.

Example: with my `control-fill-column` and
my `define-universal-argument-space-keys`:
if I hit "79 | |", it would set the column
and toggle off the indicator, but if I hit
" 79| |" (space before digits, which flows
more naturally for me most of the time),
the toggle-off would fail. With this commit
in place, the latter correctly works like
the former.

I suspect `real-{this,last}-command` should
also be faked, since this isn't normal
superficial pretending but a deeper emulation
of the Emacs command/event loop, but I'll
wait for more experience to verify that.