.emacs.el: count-visual-lines: work around point/cursor getting stuck
Example:

1. when using CRM in Vertico,

2. when I hit my enhanced `,` keybind to both
   complete the current candidate and put a
   trailing separator in the input (invalid
   character with `display` property to look
   like a comma with special glyph face),

3. if the resultant input wraps with
   just the separator on the last line,

then `end-of-visual-line` never advances
from before that last character (displayed
as point at start of last visual line) to
the end of the last line.

The cause is not particularly clear, but this
commit simply works around it by detecting the
"point is stuck (hasn't moved)" situation and
forcing the point one character forward, which
seems to work for now (and if it ever fails,
that has good odds of clarifying what's wrong).