.emacs.el: Vertico: disable truncation; resize accordingly
Finally, the Vertico experience I always
wanted: a lot of the time, if a candidate is
truncated, I want to see all of it. Especially
if something is being matched in the truncated
part.

This is less of an issue with Consult
commands (where the preview lets us see
the full matched candidate outside of the
minibuffer), but even then, my eyes are
already on the candidate list a lot of
the time - the ideal experience is both.

(I also wouldn't mind a variant where the
candidate space remained at ten lines and
untruncated line-wrap caused less candidates to
be shown, but that requires a more complicated
interaction with Vertico's logic - adjusting
`vertico-count` or similar, so that Vertico
doesn't think the off-screen candidates can
still be selected without scrolling/changing
the overlay contents.)