.emacs.el: fix over-estimate in my Vertico resize
We only need to insert the candidate overlay's
contents, because it displays after all of the
buffer text. The count overlay displays before
buffer text, so line motions won't undercount
it (unless the count overlay spans multiple
lines - but that has never come up and probably
never will, so not a priority for now).