.emacs.el: fixed-field-at-position
An attempt at a better `field-at-pos` which
actually fully takes stickiness into account.

In particular, it will correctly detect the
case where position is (on) the edge of two
adjacent fields, but isn't "in" either of
them (due to non-stickiness on both edges).

But I am leaving out any special
handling of `boundary` fields for now.