.emacs.el: smoother-fill-paragraph: recursive approach
This is just an inefficient and incomplete
proof-of-concept. It mostly makes paragraphs
look nicer, and for the cases it covers it's
a lot closer to my manual practice. Some
paragraphs are worse, it perceptibly lags for
longer paragraphs due to the recursion, and
the cases it doesn't cover can't be handled
by simple recursion (and even then only
with a quadratic time complexity increase).