.emacs.el: parent-buffer & child-buffers
Generalizing/unifying a concept that occurs
endlessly throughout Emacs: buffers that are
linked to some other buffer (either to act
on it or to provide information about it).

Examples: my partial save/revert, all of my
file-specific Git buffers, Undo Tree (note:
Undo Tree diff is the child of Undo Tree
visualizer, but code interested in acting
on the current file would want to recursively
look up the buffer-file-name of the parent of
the visualizer).