.emacs.el: android-trash
Basic building blocks for integrating
with Android trash - or rather, trashing
and restoring files following the same
spec/format/convention that Android uses.

For me, the main value of Android's recycle
bin approach (vs other systems') is that
it's sync-friendly: if I trash a file on
one device in a synced folder, it's also
trashed - not deleted - on synced devices,
and I can restore it from any device.