.emacs.el: pop-to-command: drop buffer name prefix
No longer include the "type" of command pop.

After over a year of usage, the prefix has
approximately never been useful. (Conversely,
the "eshell: " prefix has undesirably matched
a decent amount of times.)

In day-to-day use, I only ever use the Eshell
variant. The Eat variant has been invoked a
couple times for testing - and it *is* nice
for quickly verifying if some behavior comes
from an Eshell quirk or something more common
or idiomatic in Emacs.

The same information (is this popped command
using Eshell or Eat?) is available in the
modeline anyway. (I originally thought maybe
I'd want to have it in the buffer name for
minibuffer completion and such, but per prior
points, that hasn't come up.)