.emacs.el: make yt-dlp behave less wrongly
The only correct default behavior is: if a link
points to a video and merely also references a
playlist, download the video; if a link points
to playlist and merely also references a video,
download the playlist.

There is no world where the correct behavior
only considers the presence of references
without considering the semantics of which
reference (video or playlist) is primary.
Am I looking at a video (from/in a playlist)
or at a playlist (at a video)?

Sadly, `yt-dlp` is doubly-wrong here: not
only does it do the wrong thing by default,
it has no option to automatically do the
right thing - the only choice is which side
to err on.

Luckily, I have never in my life wanted to
download an entire playlist, so for my
purposes, `--no-playlist` is good enough
here.

Actually, okay, I can see an argument for
`yt-dlp`'s design being correct here:

1. if I have two distinct key binds/sequences,
   one for when I intend to grab a single
   video and one for when I intend to get the
   whole playlist, then that's a better UX,
   because these two distinct mental movements
   get mapped to separate physical movements,

2. IF the above is so much better than a DWIM
   guess that most people should be steered
   towards a separate-expressions-of-intent
   approach, then it's good for for `yt-dlp`
   to omit this logic.

But since `yt-dlp` already has to UNDERSTAND
the links, it's far far far better factoring
for the world as a whole for that extra bit
of understanding and decision-making to be
in `yt-dlp` rather than implemented outside
of it at significant duplication cost.

(UNLESS of course yt-dlp exposes most of
its site-specific link-understanding as a
library (which doesn't pull in any
networking/protocol/download dependencies
at all) - then it would be reasonable and
in fact even better for the "is this a
video link (referencing a playlist) or a
playlist link (referencing a video)?"
logic to exist separately, so that code
besides `yt-dlp` can benefit too. BUT
even in that case, I'd be tempted to say
that a tool like.`yt-dlp` might as well
include that too (at least by default).)