.emacs.el: parent-buffer: fix cleanup bug
Suppose we have three buffers, in a parent
-> child -> grandchild relationship:

    b1 -> b2 -> b3

(For example b1 might be `partial-save`, b2
might be `git add -p`, and b3 might be `git
commit`. Typical workflow for me.)

This commit fixes a bug that happens if the
buffers are killed in the order b2, b1, b3.

b2's kill hook

1. looks at b2's parent, sees b1, goes to b1
   and removes b2 from the b1's child list,
   and then

2. for all children of b2, sees b3, and sets
   b3's parent link to b2's parent (b1). So now
   b1 no longer records b2 as a child, and b3
   records b1 as its parent. BUT I forgot to
   update b1's child list to contain b3. (The
   code was correctly written during an earlier
   draft, where the `child-buffers` list didn't
   exist, and I forgot to update it.)

So that digs the pitfall.

Now b1's kill hook doesn't see a record
for b3 in b1's child list, so it doesn't
know to update b3's parent link.

So when b3's kill hook runs, it still has a
parent link to b1, and it tries to update
b1's `child-buffers` to remove itself, but
b1 has already been killed, causing an error
about selecting a killed buffer.