.emacs.el: update-dired-{file,entry}: fix parameter naming/usage
My thinking here was that semantically, by the
time we're updating, there is a current/actual
path name, and possibly a different old name.

The relevant factor is that we're being called
after a change, not before (a rename function
is called before, so it has a current/actual
path and a new intended path).

It looks like I just forgot to actually finish
implementing that idea.