.emacs.el: evil-dired-delete: don't move point
On the one hand, not moving point is more
consistent with vi in a concrete positional
way. On the other hand, moving point down
past all now-trashed paths is consistent
with vi in an abstract conceptual way: `d`
results in point being on the first thing
after the deleted items.

The bigger thing is practical - since `d`
trashes non-trash files and deletes trash
files, it's more helpful and intuitive to
remain in place, so that you can just repeat
a `d{{motion}}` (or just `.`) to immediately
delete one or more files you just trashed.

(Automatic move is also uncomfortable and
unnatural, much like a vehicle listing/pulling
to one side. I often found that frustrating
with how Dired's mark/unmark commands work. I
want to go bottom-up or just stay on the same
file just as often as I want to go top-down.)