.emacs.el: next-string-split-separator
Like `next-single-property-change`, but for
`string-split` boundaries rather than field
boundaries.

* Since it works on buffers as well as
  strings, it's a good core for using
  string-split semantics for buffer contents.

* Since it returns a position within a buffer
  or string rather than returning already split
  strings, it makes it easy to combine several
  different kinds of boundaries - for example,
  looking for either the next field boundary or
  the next newline, whichever is encountered
  first.

The `position`, `object`, and `limit`
parameters behave the same as in
`next-single-property-change`.

The `separators` and `omit-nulls` parameters
behace the same as in `string-split`.

The position returned is the start of the
next separator that `string-split` would
split at. If there are no more separators,
returns `limit`, or nil if there is no limit.

Since string-split can use an arbitrary regular
expression for separators, this function also
sets match data to the start and end of the
separator (the entire span between the two
substrings that `string-split` would produce).

Since `omit-nulls` effectively treats multiple
adjacent separators as one big separator, the
match data will be at the end of the last
separator.

In other words, `match-end` always points
at the start of the string that would be
produced by an equivalent `split-string`
(without any `trim` argument).

There is no special handling for the very
start and end - in particular, a leading
separator with nothing before it, or a
trailing separator with nothing after it,
needs to be detected separately. I don't
see a great way for this "start of next
boundary" API to automatically handle
both in a consistenr way.