.emacs.el: full-path-property-split: support partially propertized text
Handle mixed inputs: some un-propertized lines
(such as paths typed manually), and some text
annotated with my `full-path` property.

Prior to this commit, the presence of the
`full-path` property would cause only those
property values to be used. This is unlikely
to come up in practice, but it would behave
wrongly (unexpectedly ignore some files) if a
file list was built up from a mix of sources.