.emacs.el: update-dired-{entry,file}: handle moves
If the file moves from one directory to another
rather than (just) being renamed, remove its
entries from Dired buffers of the source
directory, and add its entries to any Dired
buffers of the destination.

This behavior is much more convenient DX than
requiring the caller to choose between update
or an add-remove pair.

Notably, when not moving between directories,
the entry is still only updated if it was
already there. This is intentional, because if
the entry was intentionally removed/hidden, it
should stay removed/hidden.