.emacs.el: evil-dired-replacing-paste
This completes my vi file management `p` and
`P` with their analogs of my `gp` and `gP`.

Ideally we'd pass the `ok-if-already-exists`
flag and let the underlying functions do the
overwrite (often atomically, since `rename(2)`
is the underlying primitive), but that fails
when replacing a directory, so I went with
this commit's simpler method.