.emacs.el: full-path-property: Dired: raw path
Use the expanded/unabbreviated path, without
a trailing slash to indicate directory name.
For example, "/home/user/Downloads" instead
of "~/Downloads/".

I have some mixed feelings about the merit
of this, especially losing the trailing
slash (in particular, I suspect a formal
spec for `full-path` would tell consumers
that they MUST handle both abbreviated and
not, both with trailing slash and without),
but it's simpler and consistent with how my
Music buffer and Eshell integration do it.

Increasingly, if I want to see the path
in a human-friendly form, I just call my
`user-friendly-path` on it anyway, rather
than just the built-in abbreviation. (And
I don't want to make `full-path-property`
depend on `user-friendly-path`.)