.emacs.el: fixed-dired-omit-expunge
"(Nothing to omit)" is noise unless the
user has just taken an action that is
expected/intended to omit some files.

But just flipping `dired-omit-verbose` would be
wrong, since the other messages ought to show
up whenever a new file appearance (which might
be expected/intended) has been suppressed.

Ideally, we'd advise `dired-omit-expunge`
to advise `message` to do nothing if its
argument is "(Nothing to omit)", but
`dired-omit-expunge` is compiled and
`message` is a primitive, so
`dired-omit-expunge` would call the
unadvised `message` directly through C.
Next best thing I found was to advise
`dired-omit-expunge` to update
`set-message-functions`.