bin/git: stop setting GIT_CEILING_DIRECTORIES
After a year or two of experimenting with
this, I've grown tired of the downside
(worse UX for directories that *are* meant
to be version-controlled as part of my
home directory, such as ~/.config (and,
at least for the time being, ~/bin)).

I've finally figured out what seems like
a better method: add a blocker `.git` file
with the contents

    gitdir: /dev/null

to subdirectories as needed.

This allows still having a Git-managed
home directory while keeping all the
benefits of having the standard gitdir
location and preventing Git commands in
(some) subdirectories from considering
themselves part of the home repo. We
gain the granularity of letting some
directories through the ceiling, at the
cost of needing to copy the `.git`
blocker file to multiple directories.

In practice, I've found that the granularity
of per-directory choice is desirable.

Also in practice, I've found that
approximately all Git re-implementations
fail to support `GIT_CEILING_DIRECTORIES`
correctly. (Conversely, both Exa and Emacs'
`vc` understand the common case of looking
rootwards in the file system for ".git".)

As a result of the above, I also no longer
think it's right for the `git` command to
automatically treat the home directory as
a ceiling directory.