.emacs.el: evil-dired-show-hidden; hide dotfiles
I always wanted my file managers to be at
least as nice as `ls` and `ls -a`: hide
hidden files by default, but let me see
them on a case-by-case basis with just a
couple keystrokes.

Since hiding/unhiding files can cause the
lines to shift around on the screen, I pulse
the current line after the refresh, to make
to visually easier to find where the cursor
(and the current file, if it's still visible)
went.