.emacs.el: evil-dired-delete: delete trashed lines
This was the first approach I imagined:
automatically delete the lines of the
directory entries that were trashed by
this delete operation. (In a mature
public release, this would be optional.)

This allows viewing trashed files with a mere
revert or toggle of "show hidden files", but
maintains the vi-style UX of lines actually
being deleted in the moment.