Revert ".emacs.el: evil-dired-delete: bypass trash with _"
I don't actually want to use this for now. The
black hole register in vi is not semantically a
full-blown "shred" operation (deleting to it is
not overwriting the bytes in memory and leaving
a placeholder in the undo history that prevents
retrieval of the original value - it only
bypasses the clipboard/registers/kill-ring).

This reverts commit fa36754be5a8a03e1e65bc61bd24cb236d40014b.