.emacs.el: make ace-window not overpad hints
When showing a window hint over a character
that's wider than one column (for example,
a tab), ace-window pads the window hint.

But the code that adds the padding
assumes the hint is only one character
wide, even though window hints can be
more than one character (for example,
when `aw-leading-char-style` is set to
`path`, the tenth window gets hint "91").

This commit partially compensates for this
problem, in a way that will work as long
as the hint isn't wider than the overlaid
character.