.emacs.el: make dired-revert scroll almost perfect
This is one of the bigger issues I have with
Dired's revert: in a large directory, reverting
Dired causes the window to scroll so that the
current file ends up jumping to the bottom.

This commit provides a first-pass fix of
that, by advising `dired-save-positions` to
save not just the filename and buffer line,
but also the visual line within the window,
and advising `dired-revert-positions` to
scroll/recenter the window to match.

(This code also absorbs my prior advice to
`dired-revert` to preserve column, because
restoring the column is necessarily coupled to
restoring the scroll: if the column is large
enough to be on the continuation line of a
line-wrapped buffer line, and the start of that
line would be above the top of the window, then
we need to restore the column first, so that
point doesn't hold back the scrolling. Also,
ideally this code is eventually generalized to
other windows displaying the same buffer, and
in windows' entries of previously displayed
buffers, and then this is the best place for
restoring all of those columns. Finally, the
{save,restore}-position functioms are
semantically the right place for it, since
column is a part/type of position.)

The reason for using negative line index from
the bottom of window for `recenter` is that
it seems to bypass a weird bug that seems to
afflict `recenter`, where a positive line
offset sometimes results in the window being
scrolled more lines than necessary (I've only
seen this bug in Dired, when there are
line-wrapped lines spanning the desired
window start position, but it doesn't happen
in all such cases, and I haven't yet found
the exact conditions for this bug - but it
seems that in all cases where it happens,
passing an equivalent negative argument to
`recenter` avoids the problem.

Since a pure scroll restore can result in
under-utilization of the window (for example,
when many entries are added by the revert,
and the Dired buffer goes from partially
filling the window to totally filling the
window), besides simply restoring the scroll
position, we also scroll to fill the window
if there are lines to show out-of-view above
and unused lines at the bottom.